refactor: replace PlantContext with PlantContextHelper
- Deleted the PlantContext.ts file and updated imports in PlantContext.tsx and usePlants.ts to reference PlantContextHelper instead.
- This change streamlines the context management for plant data.
