--- leafDex_Web/src/context/PlantContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-import { PlantContextType } from '../types/plants';
-
-export const PlantContext = createContext<PlantContextType | undefined>(
-  undefined
-);
